Added a relationship between listing and the user && made sure only the specific user can edit/delete his/her listing

+++ app/Http/Controllers/ListingController.php
@@ -59,6 +59,12 @@ class ListingController extends Controller
     //Update Listing
     public function update(Request $request, Listing $listing){
 
+        //Make sure that logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'unauthorized action');
+        }
+
+        //Validate formFields while updating
         $formFields = $request->validate([
             'title' =>'required',
             'company' => 'required',
@@ -78,11 +84,19 @@ class ListingController extends Controller
     //Delete Listing
 
     public function destroy(Listing $listing){
+
+        //Make sure that logged in user is the owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'unauthorized action');
+        } 
+
+        //Request to destroy a listing 
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
 
+    //Manage Listings
     public function manage(){
         return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
